Modified DisposeOfRenderer to be called DestroyRenderer.  Moved significantly sized interface implementations into separate files via partial classes.
